requirements.txt
asgiref==3.3.4
boto3==1.18.22
botocore==1.21.22
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.8.0
django-filter==2.4.0
django-heroku==0.3.1
django-storages==1.11.1
dnspython==1.16.0
psycopg2==2.8.6
pymongo==3.12.0
python-decouple==3.4
docutils==0.17
gunicorn==20.1.0
pytz==2021.1
s3transfer==0.5.0
six==1.15.0
urllib3==1.26.4
whitenoise==5.3.0

pyPdf2==1.26.0 
python-bidi==0.4.2 
reportlab==3.6.2 
xhtml2pdf==0.2.5

